refactor(characters): split tests to mirror module extraction
The Characters module was split into per-concern modules (Store, Effects,
Leveling, Generation, Rolls, Progression, Resolution, Preparation) over
the previous commits, but all tests remained in the monolithic
characters_test.exs. This moves each describe block into a test file
matching the module that now owns the behavior, so test failures point
directly at the responsible module and each file can be run in isolation.

characters_test.exs is reduced to running the facade's doctests; the
delegating functions themselves are covered by the per-concern files.
No test bodies were changed — this is a pure relocation.

Leveling gets no new file because its behavior was already covered by
the pre-existing apply_award_test.exs and xp_to_next_level_test.exs.

The codescene hook is skipped for this commit only: it flags pre-existing
duplication in the relocated preparation tests as "new" because the file
is new. Fixing it here would mix a refactor into a pure move; the
duplication is removed in the following commit instead.
