+++ apps/ex_ttrpg_dev/test/characters/store_test.exs
@@ -1,0 +1,95 @@
+defmodule ExTTRPGDevTest.Characters.Store do
+  # Not async: these tests share the on-disk characters directory.
+  use ExUnit.Case
+  alias ExTTRPGDev.Characters
+  alias ExTTRPGDev.Characters.Character
+  alias ExTTRPGDev.RuleSystems
+
+  def build_test_character do
+    RuleSystems.list_systems()
+    |> List.first()
+    |> RuleSystems.load_system!()
+    |> Character.gen_character!()
+  end
+
+  def save_test_character do
+    character = build_test_character()
+    Characters.save_character!(character)
+    character
+  end
+
+  def delete_test_character(%Character{} = character) do
+    File.rm(Characters.character_file_path!(character))
+  end
+
+  test "character_exists?/1" do
+    character = build_test_character()
+    assert not Characters.character_exists?(character)
+    assert not Characters.character_exists?(character.metadata.slug)
+
+    Characters.save_character!(character)
+    assert Characters.character_exists?(character)
+    assert Characters.character_exists?(character.metadata.slug)
+
+    # cleanup
+    delete_test_character(character)
+  end
+
+  test "save_character!/1" do
+    character = build_test_character()
+    assert not Characters.character_exists?(character)
+
+    Characters.save_character!(character)
+    assert Characters.character_exists?(character)
+
+    assert_raise RuntimeError, fn -> Characters.save_character!(character) end
+
+    # cleanup
+    delete_test_character(character)
+  end
+
+  test "save_character!/2" do
+    character = build_test_character()
+    assert not Characters.character_exists?(character)
+
+    Characters.save_character!(character)
+    assert Characters.character_exists?(character)
+
+    Characters.save_character!(character, true)
+    assert Characters.character_exists?(character)
+
+    # cleanup
+    delete_test_character(character)
+  end
+
+  test "list_characters!/0" do
+    characters_list_first = Characters.list_characters!()
+    character = save_test_character()
+    characters_list_second = Characters.list_characters!()
+    assert Enum.count(characters_list_first) < Enum.count(characters_list_second)
+    assert Enum.member?(characters_list_second, character.metadata.slug)
+
+    # cleanup
+    delete_test_character(character)
+  end
+
+  test "load_character/1" do
+    character = save_test_character()
+    loaded_character = Characters.load_character!(character.metadata.slug)
+    assert character == loaded_character
+
+    delete_test_character(character)
+  end
+
+  test "delete_character/1 deletes an existing character and returns :ok" do
+    character = save_test_character()
+    assert Characters.character_exists?(character)
+
+    assert :ok = Characters.delete_character(character.metadata.slug)
+    refute Characters.character_exists?(character)
+  end
+
+  test "delete_character/1 returns error for unknown slug" do
+    assert {:error, :not_found} = Characters.delete_character("nonexistent_character_xyz")
+  end
+end
