+++ apps/ex_ttrpg_dev/test/characters_test.exs
@@ -1,12 +1,8 @@
 defmodule ExTTRPGDevTest.Characters do
-  use ExUnit.Case
-  alias ExTTRPGDev.Characters
-  alias ExTTRPGDev.Characters.{Character, Decision, InventoryItem}
-  alias ExTTRPGDev.RuleSystem.Effect
-  alias ExTTRPGDev.RuleSystem.InventoryRules
-  alias ExTTRPGDev.RuleSystem.Node
-  alias ExTTRPGDev.RuleSystems
-  alias ExTTRPGDev.RuleSystems.LoadedSystem
+  # The Characters module is a thin facade; its behavior is tested per
+  # concern in test/characters/*_test.exs. This module only runs the
+  # facade's doctests.
+  use ExUnit.Case, async: true
 
   doctest ExTTRPGDev.Characters,
     except: [
@@ -18,1617 +14,4 @@ defmodule ExTTRPGDevTest.Characters do
       load_character!: 1,
       concept_roll!: 4
     ]
-
-  def build_test_character do
-    RuleSystems.list_systems()
-    |> List.first()
-    |> RuleSystems.load_system!()
-    |> Character.gen_character!()
-  end
-
-  def save_test_character do
-    character = build_test_character()
-    Characters.save_character!(character)
-    character
-  end
-
-  def delete_test_character(%Character{} = character) do
-    File.rm(Characters.character_file_path!(character))
-  end
-
-  test "character_exists?/1" do
-    character = build_test_character()
-    assert not Characters.character_exists?(character)
-    assert not Characters.character_exists?(character.metadata.slug)
-
-    Characters.save_character!(character)
-    assert Characters.character_exists?(character)
-    assert Characters.character_exists?(character.metadata.slug)
-
-    # cleanup
-    delete_test_character(character)
-  end
-
-  test "save_character!/1" do
-    character = build_test_character()
-    assert not Characters.character_exists?(character)
-
-    Characters.save_character!(character)
-    assert Characters.character_exists?(character)
-
-    assert_raise RuntimeError, fn -> Characters.save_character!(character) end
-
-    # cleanup
-    delete_test_character(character)
-  end
-
-  test "save_character!/2" do
-    character = build_test_character()
-    assert not Characters.character_exists?(character)
-
-    Characters.save_character!(character)
-    assert Characters.character_exists?(character)
-
-    Characters.save_character!(character, true)
-    assert Characters.character_exists?(character)
-
-    # cleanup
-    delete_test_character(character)
-  end
-
-  test "list_characters!/0" do
-    characters_list_first = Characters.list_characters!()
-    character = save_test_character()
-    characters_list_second = Characters.list_characters!()
-    assert Enum.count(characters_list_first) < Enum.count(characters_list_second)
-    assert Enum.member?(characters_list_second, character.metadata.slug)
-
-    # cleanup
-    delete_test_character(character)
-  end
-
-  test "load_character/1" do
-    character = save_test_character()
-    loaded_character = Characters.load_character!(character.metadata.slug)
-    assert character == loaded_character
-
-    delete_test_character(character)
-  end
-
-  test "delete_character/1 deletes an existing character and returns :ok" do
-    character = save_test_character()
-    assert Characters.character_exists?(character)
-
-    assert :ok = Characters.delete_character(character.metadata.slug)
-    refute Characters.character_exists?(character)
-  end
-
-  test "delete_character/1 returns error for unknown slug" do
-    assert {:error, :not_found} = Characters.delete_character("nonexistent_character_xyz")
-  end
-
-  describe "active_concepts/2" do
-    test "returns empty set for no decisions" do
-      assert MapSet.new() == Characters.active_concepts([], %{})
-    end
-
-    test "returns root concept for a single root decision" do
-      decisions = [%Decision{scope: nil, choice: "race", selection: "human"}]
-      result = Characters.active_concepts(decisions, %{})
-      assert MapSet.member?(result, {"race", "human"})
-    end
-
-    @dwarf_metadata %{
-      {"race", "dwarf"} => %{
-        "choices" => %{"subrace" => %{"type" => "race", "options" => ["hill_dwarf"]}}
-      }
-    }
-
-    test "recurses into sub-choices when a decision exists for them" do
-      decisions = [
-        %Decision{scope: nil, choice: "race", selection: "dwarf"},
-        %Decision{scope: {"race", "dwarf"}, choice: "subrace", selection: "hill_dwarf"}
-      ]
-
-      result = Characters.active_concepts(decisions, @dwarf_metadata)
-      assert MapSet.member?(result, {"race", "dwarf"})
-      assert MapSet.member?(result, {"race", "hill_dwarf"})
-    end
-
-    test "does not activate sub-concept when no decision is made for a choice" do
-      decisions = [%Decision{scope: nil, choice: "race", selection: "dwarf"}]
-      result = Characters.active_concepts(decisions, @dwarf_metadata)
-      assert MapSet.member?(result, {"race", "dwarf"})
-      refute MapSet.member?(result, {"race", "hill_dwarf"})
-    end
-
-    @fighter_with_inventory_choice %{
-      {"class", "fighter"} => %{
-        "choices" => %{
-          "starting_armor" => %{
-            "type" => "equipment",
-            "grants_to" => "inventory",
-            "options" => ["chain_mail"]
-          }
-        }
-      }
-    }
-
-    test "does not recurse into choices with grants_to: inventory" do
-      decisions = [
-        %Decision{scope: nil, choice: "class", selection: "fighter"},
-        %Decision{scope: {"class", "fighter"}, choice: "starting_armor", selection: "chain_mail"}
-      ]
-
-      result = Characters.active_concepts(decisions, @fighter_with_inventory_choice)
-      assert MapSet.member?(result, {"class", "fighter"})
-      refute MapSet.member?(result, {"equipment", "chain_mail"})
-    end
-  end
-
-  describe "random_decisions/1" do
-    setup do
-      {:ok, system: RuleSystems.load_system!("dnd_5e_srd")}
-    end
-
-    test "returns one root decision per character choice", %{system: system} do
-      decisions = Characters.random_decisions(system)
-      root = Enum.filter(decisions, &(&1.scope == nil))
-      assert length(root) == length(system.module.character_building_choices)
-    end
-
-    test "root decision choice matches the character_choice concept_type", %{system: system} do
-      decisions = Characters.random_decisions(system)
-
-      for %{concept_type: type_id} <- system.module.character_building_choices do
-        assert Enum.any?(decisions, &(&1.scope == nil and &1.choice == type_id))
-      end
-    end
-
-    test "selected root race is not a subrace", %{system: system} do
-      decisions = Characters.random_decisions(system)
-      root_race = Enum.find(decisions, &(&1.scope == nil and &1.choice == "race"))
-
-      subraces = ~w[hill_dwarf high_elf lightfoot_halfling rock_gnome]
-
-      refute root_race.selection in subraces
-    end
-
-    test "equipment choices (grants_to: inventory) produce a decision but do not recurse", %{
-      system: system
-    } do
-      concept_metadata =
-        Map.put(system.concept_metadata, {"class", "fighter"}, %{
-          "choices" => %{
-            "starting_weapon" => %{
-              "type" => "equipment",
-              "grants_to" => "inventory",
-              "options" => ["longsword", "shortsword"]
-            }
-          }
-        })
-
-      system = %{system | concept_metadata: concept_metadata}
-
-      for _ <- 1..10 do
-        decisions = Characters.random_decisions(system)
-
-        weapon_decision =
-          Enum.find(decisions, fn d ->
-            d.scope != nil and elem(d.scope, 0) == "class" and d.choice == "starting_weapon"
-          end)
-
-        if weapon_decision do
-          assert weapon_decision.selection in ["longsword", "shortsword"]
-          # The selected equipment id should not appear as a scope in any decision
-          refute Enum.any?(decisions, fn d ->
-                   d.scope != nil and elem(d.scope, 1) == weapon_decision.selection
-                 end)
-        end
-      end
-    end
-
-    test "races with subraces produce a subrace decision", %{system: system} do
-      races_with_subraces = ~w[dwarf elf halfling gnome]
-
-      for _ <- 1..20 do
-        decisions = Characters.random_decisions(system)
-        root_race = Enum.find(decisions, &(&1.scope == nil and &1.choice == "race"))
-
-        if root_race.selection in races_with_subraces do
-          parent_scope = {"race", root_race.selection}
-          assert Enum.any?(decisions, &(&1.scope == parent_scope and &1.choice == "subrace"))
-        end
-      end
-    end
-  end
-
-  describe "active_effects/2" do
-    @fighter_effect %Effect{
-      source: {"class", "fighter"},
-      target: {"saving_throw", "strength", "modifier"},
-      value: 2
-    }
-
-    defp minimal_system(effects, concept_metadata) do
-      %ExTTRPGDev.RuleSystems.LoadedSystem{
-        module: nil,
-        graph: nil,
-        nodes: %{},
-        rolling_methods: %{},
-        effects: effects,
-        concept_metadata: concept_metadata
-      }
-    end
-
-    defp minimal_character(decisions) do
-      %Character{
-        name: "Test",
-        generated_values: %{},
-        effects: [],
-        decisions: decisions,
-        metadata: %ExTTRPGDev.Characters.Metadata{slug: "test", rule_system: "dnd_5e_srd"}
-      }
-    end
-
-    test "includes system effects for active concepts" do
-      system = minimal_system([@fighter_effect], %{{"class", "fighter"} => %{}})
-
-      character =
-        minimal_character([%Decision{scope: nil, choice: "class", selection: "fighter"}])
-
-      effects = Characters.active_effects(system, character)
-      assert Enum.any?(effects, &(&1.source == {"class", "fighter"}))
-    end
-
-    test "excludes system effects for inactive concepts" do
-      system = minimal_system([@fighter_effect], %{})
-      character = minimal_character([])
-      effects = Characters.active_effects(system, character)
-      refute Enum.any?(effects, &(&1.source == {"class", "fighter"}))
-    end
-
-    test "generates effects from decisions with contributes_field" do
-      concept_metadata = %{
-        {"class", "fighter"} => %{
-          "choices" => %{
-            "skill_proficiency_1" => %{
-              "type" => "skill",
-              "contributes_field" => "modifier",
-              "contributes_value" => "character_trait('proficiency_bonus').bonus",
-              "options" => ["athletics", "acrobatics"]
-            }
-          }
-        }
-      }
-
-      system = minimal_system([], concept_metadata)
-
-      character =
-        minimal_character([
-          %Decision{
-            scope: {"class", "fighter"},
-            choice: "skill_proficiency_1",
-            selection: "athletics"
-          }
-        ])
-
-      effects = Characters.active_effects(system, character)
-
-      assert Enum.any?(
-               effects,
-               &(&1.source == {"class", "fighter"} and
-                   &1.target == {"skill", "athletics", "modifier"} and
-                   &1.value == "character_trait('proficiency_bonus').bonus")
-             )
-    end
-
-    test "includes effects from inventory items with item_fields populated" do
-      system =
-        minimal_system(
-          [
-            %Effect{
-              source: {"equipment", "longsword"},
-              target: {"character_trait", "ac", "value"},
-              value: 2,
-              when: "item.equipped"
-            }
-          ],
-          %{}
-        )
-
-      item = %InventoryItem{
-        concept_type: "equipment",
-        concept_id: "longsword",
-        fields: %{"equipped" => true}
-      }
-
-      character = %{minimal_character([]) | inventory: [item]}
-      effects = Characters.active_effects(system, character)
-
-      assert length(effects) == 1
-      [effect] = effects
-      assert effect.source == {"equipment", "longsword"}
-      assert effect.item_fields == %{"equipped" => true}
-    end
-
-    test "produces one effect entry per inventory item even for the same concept" do
-      system =
-        minimal_system(
-          [
-            %Effect{
-              source: {"equipment", "shortsword"},
-              target: {"stat", "atk", "bonus"},
-              value: 1
-            }
-          ],
-          %{}
-        )
-
-      item1 = %InventoryItem{concept_type: "equipment", concept_id: "shortsword", fields: %{}}
-      item2 = %InventoryItem{concept_type: "equipment", concept_id: "shortsword", fields: %{}}
-
-      character = %{minimal_character([]) | inventory: [item1, item2]}
-      effects = Characters.active_effects(system, character)
-      assert length(effects) == 2
-    end
-
-    test "does not generate effects for choices without contributes_field" do
-      concept_metadata = %{
-        {"race", "elf"} => %{
-          "choices" => %{
-            "subrace" => %{"type" => "race", "options" => ["high_elf"]}
-          }
-        }
-      }
-
-      system = minimal_system([], concept_metadata)
-
-      character =
-        minimal_character([
-          %Decision{scope: {"race", "elf"}, choice: "subrace", selection: "high_elf"}
-        ])
-
-      assert Characters.active_effects(system, character) == []
-    end
-  end
-
-  describe "pending_choices/3" do
-    @level_binding %{{"character_trait", "character_level", "level"} => 3}
-
-    @hp_progression %{
-      "name" => "Hit Points",
-      "required_count" => "character_trait('character_level').level - 1",
-      "effect_target" => "character_trait('max_hit_points').points"
-    }
-
-    @spend_xp_progression %{
-      "name" => "Spend XP",
-      "available_when" => "character_trait('experience_points').total",
-      "effect_target" => "skill('athletics').modifier"
-    }
-
-    defp progression_system(progressions) do
-      minimal_system(
-        [],
-        Map.new(progressions, fn {id, meta} ->
-          {{"character_progression", id}, meta}
-        end)
-      )
-    end
-
-    test "returns empty list when no character_progression concepts exist" do
-      system = minimal_system([], %{})
-      assert Characters.pending_choices(system, minimal_character([]), %{}) == []
-    end
-
-    test "returns a pending entry when required slots exceed decisions made" do
-      system = progression_system(%{"hp_per_level" => @hp_progression})
-      # level 3 → required 2, decisions made 0 → 2 pending
-      [entry] = Characters.pending_choices(system, minimal_character([]), @level_binding)
-      assert entry.type == :pending
-      assert entry.id == "hp_per_level"
-      assert entry.name == "Hit Points"
-      assert entry.count == 2
-      assert entry.effect_target == "character_trait('max_hit_points').points"
-    end
-
-    test "reduces pending count by decisions already made for that progression" do
-      system = progression_system(%{"hp_per_level" => @hp_progression})
-
-      decisions = [
-        %Decision{
-          scope: {"character_progression", "hp_per_level"},
-          choice: "choice_1",
-          selection: "rolled"
-        }
-      ]
-
-      # level 3 → required 2, made 1 → 1 pending
-      [entry] = Characters.pending_choices(system, minimal_character(decisions), @level_binding)
-      assert entry.count == 1
-    end
-
-    test "returns empty when all required slots are filled" do
-      system = progression_system(%{"hp_per_level" => @hp_progression})
-
-      decisions = [
-        %Decision{
-          scope: {"character_progression", "hp_per_level"},
-          choice: "choice_1",
-          selection: "rolled"
-        },
-        %Decision{
-          scope: {"character_progression", "hp_per_level"},
-          choice: "choice_2",
-          selection: "average"
-        }
-      ]
-
-      # level 3 → required 2, made 2 → empty
-      assert Characters.pending_choices(system, minimal_character(decisions), @level_binding) ==
-               []
-    end
-
-    test "does not count decisions scoped to a different progression" do
-      system =
-        progression_system(%{"hp_per_level" => @hp_progression, "other" => @hp_progression})
-
-      decisions = [
-        %Decision{
-          scope: {"character_progression", "other"},
-          choice: "choice_1",
-          selection: "rolled"
-        }
-      ]
-
-      result = Characters.pending_choices(system, minimal_character(decisions), @level_binding)
-      hp_entry = Enum.find(result, &(&1.id == "hp_per_level"))
-      assert hp_entry.count == 2
-    end
-
-    test "returns empty for required_count when formula binding is missing" do
-      system = progression_system(%{"hp_per_level" => @hp_progression})
-      assert Characters.pending_choices(system, minimal_character([]), %{}) == []
-    end
-
-    test "returns an available entry when available_when is truthy" do
-      system = progression_system(%{"spend_xp" => @spend_xp_progression})
-      resolved = %{{"character_trait", "experience_points", "total"} => 100}
-      [entry] = Characters.pending_choices(system, minimal_character([]), resolved)
-      assert entry.type == :available
-      assert entry.id == "spend_xp"
-      assert entry.effect_target == "skill('athletics').modifier"
-    end
-
-    test "returns empty when available_when evaluates to 0" do
-      system = progression_system(%{"spend_xp" => @spend_xp_progression})
-      resolved = %{{"character_trait", "experience_points", "total"} => 0}
-      assert Characters.pending_choices(system, minimal_character([]), resolved) == []
-    end
-
-    test "returns empty when available_when evaluates to false" do
-      progression =
-        Map.put(
-          @spend_xp_progression,
-          "available_when",
-          "character_trait('experience_points').total > 0"
-        )
-
-      system = progression_system(%{"spend_xp" => progression})
-      resolved = %{{"character_trait", "experience_points", "total"} => 0}
-      assert Characters.pending_choices(system, minimal_character([]), resolved) == []
-    end
-
-    test "returns empty for available_when when formula binding is missing" do
-      system = progression_system(%{"spend_xp" => Map.delete(@spend_xp_progression, "name")})
-      assert Characters.pending_choices(system, minimal_character([]), %{}) == []
-    end
-
-    test "returns empty when progression has neither required_count nor available_when" do
-      progression = %{
-        "name" => "Inert",
-        "effect_target" => "character_trait('max_hit_points').points"
-      }
-
-      system = progression_system(%{"inert" => progression})
-      assert Characters.pending_choices(system, minimal_character([]), %{}) == []
-    end
-
-    test "falls back to id as name when name is not present" do
-      progression = Map.delete(@hp_progression, "name")
-      system = progression_system(%{"hp_per_level" => progression})
-      [entry] = Characters.pending_choices(system, minimal_character([]), @level_binding)
-      assert entry.name == "hp_per_level"
-    end
-
-    test "resolves roll_reference from the character's active root decision" do
-      concept_metadata = %{
-        {"character_progression", "hp_per_level"} =>
-          Map.put(@hp_progression, "roll_reference", "class.hit_die"),
-        {"class", "fighter"} => %{"hit_die" => "d10"}
-      }
-
-      system = minimal_system([], concept_metadata)
-      decisions = [%Decision{scope: nil, choice: "class", selection: "fighter"}]
-      [entry] = Characters.pending_choices(system, minimal_character(decisions), @level_binding)
-      assert entry.roll == "d10"
-    end
-
-    test "roll is nil when roll_reference has no matching decision on the character" do
-      concept_metadata = %{
-        {"character_progression", "hp_per_level"} =>
-          Map.put(@hp_progression, "roll_reference", "class.hit_die")
-      }
-
-      system = minimal_system([], concept_metadata)
-      [entry] = Characters.pending_choices(system, minimal_character([]), @level_binding)
-      assert entry.roll == nil
-    end
-
-    test "roll is nil when no roll_reference is declared" do
-      system = progression_system(%{"hp_per_level" => @hp_progression})
-      [entry] = Characters.pending_choices(system, minimal_character([]), @level_binding)
-      assert entry.roll == nil
-    end
-
-    test "returns entries for multiple progressions" do
-      other = %{
-        "name" => "Other",
-        "required_count" => "character_trait('character_level').level - 1",
-        "effect_target" => "skill('athletics').modifier"
-      }
-
-      system = progression_system(%{"hp_per_level" => @hp_progression, "other" => other})
-      result = Characters.pending_choices(system, minimal_character([]), @level_binding)
-      assert Enum.map(result, & &1.id) |> Enum.sort() == ["hp_per_level", "other"]
-    end
-
-    @cantrip_progression %{
-      "name" => "Cantrip",
-      "required_count" => "2",
-      "type" => "spell",
-      "filter" => %{
-        "level" => 0,
-        "active_in" => %{"field" => "classes", "type" => "class"}
-      }
-    }
-
-    @spell_meta %{
-      {"spell", "fire_bolt"} => %{"level" => 0, "classes" => ["wizard"]},
-      {"spell", "mage_hand"} => %{"level" => 0, "classes" => ["wizard"]},
-      {"spell", "cure_wounds"} => %{"level" => 1, "classes" => ["cleric"]},
-      {"spell", "sacred_flame"} => %{"level" => 0, "classes" => ["cleric"]}
-    }
-
-    defp spell_system(progressions, spell_meta) do
-      minimal_system(
-        [],
-        Map.merge(
-          Map.new(progressions, fn {id, meta} -> {{"character_progression", id}, meta} end),
-          spell_meta
-        )
-      )
-    end
-
-    test "spell progression includes options filtered by level and active class" do
-      system = spell_system(%{"cantrips" => @cantrip_progression}, @spell_meta)
-
-      decisions = [%Decision{scope: nil, choice: "class", selection: "wizard"}]
-      [entry] = Characters.pending_choices(system, minimal_character(decisions), %{})
-
-      assert entry.id == "cantrips"
-      assert entry.count == 2
-      assert entry.effect_target == nil
-      assert entry.roll == nil
-      assert Enum.sort(entry.options) == ["fire_bolt", "mage_hand"]
-    end
-
-    test "spell progression returns no options when no class active" do
-      system = spell_system(%{"cantrips" => @cantrip_progression}, @spell_meta)
-
-      [entry] = Characters.pending_choices(system, minimal_character([]), %{})
-
-      assert entry.options == []
-    end
-
-    test "spell progression with min/max level filter" do
-      progression = %{
-        "name" => "Spell",
-        "required_count" => "2",
-        "type" => "spell",
-        "filter" => %{
-          "min_level" => 1,
-          "max_level_node" => "character_trait('max_spell_level').level",
-          "active_in" => %{"field" => "classes", "type" => "class"}
-        }
-      }
-
-      system = spell_system(%{"spells_known" => progression}, @spell_meta)
-      resolved = %{{"character_trait", "max_spell_level", "level"} => 1}
-      decisions = [%Decision{scope: nil, choice: "class", selection: "cleric"}]
-
-      [entry] = Characters.pending_choices(system, minimal_character(decisions), resolved)
-
-      assert entry.id == "spells_known"
-      assert entry.options == ["cure_wounds"]
-    end
-
-    test "spell progression uses pending_choice_slots cap instead of resolved max level" do
-      spell_meta =
-        Map.merge(@spell_meta, %{
-          {"spell", "fireball"} => %{"level" => 3, "classes" => ["cleric"]}
-        })
-
-      progression = %{
-        "name" => "Spell",
-        "required_count" => "2",
-        "type" => "spell",
-        "filter" => %{
-          "min_level" => 1,
-          "max_level_node" => "character_trait('max_spell_level').level",
-          "active_in" => %{"field" => "classes", "type" => "class"}
-        }
-      }
-
-      system = spell_system(%{"spells_known" => progression}, spell_meta)
-      # resolved says max_spell_level = 3, but the slot was earned when only level 1 was available
-      resolved = %{{"character_trait", "max_spell_level", "level"} => 3}
-      decisions = [%Decision{scope: nil, choice: "class", selection: "cleric"}]
-
-      pending_choice_slots = [
-        %{progression_id: "spells_known", earned_at_level: 1, max_level_cap: 1}
-      ]
-
-      character = %{minimal_character(decisions) | pending_choice_slots: pending_choice_slots}
-      [entry] = Characters.pending_choices(system, character, resolved)
-
-      # fireball (level 3) should be excluded despite resolved allowing level 3
-      assert entry.options == ["cure_wounds"]
-    end
-  end
-
-  describe "compute_pending_choice_slots/2" do
-    setup do
-      system = RuleSystems.load_system!("dnd_5e_srd")
-      attrs = ~w[strength dexterity constitution wisdom intelligence charisma]
-      generated = Map.new(attrs, &{{"ability", &1, "base_score"}, 10})
-
-      decisions = [
-        %Decision{scope: nil, choice: "class", selection: "sorcerer"},
-        %Decision{scope: nil, choice: "race", selection: "human"},
-        %Decision{scope: nil, choice: "background", selection: "acolyte"}
-      ]
-
-      character = %Character{
-        name: "Test",
-        generated_values: generated,
-        effects: [],
-        decisions: decisions,
-        pending_choice_slots: [],
-        metadata: %ExTTRPGDev.Characters.Metadata{slug: "test_slots", rule_system: "dnd_5e_srd"}
-      }
-
-      {:ok, system: system, character: character}
-    end
-
-    test "level 1 character gets slots with cap 1 for initial spell choices",
-         %{system: system, character: character} do
-      slots = Characters.compute_pending_choice_slots(system, character)
-      spells_known = Enum.filter(slots, &(&1.progression_id == "spells_known"))
-
-      # sorcerer starts with 2 spells known at level 1
-      assert length(spells_known) == 2
-      assert Enum.all?(spells_known, &(&1.earned_at_level == 1))
-      assert Enum.all?(spells_known, &(&1.max_level_cap == 1))
-    end
-
-    test "higher-level slots get appropriate caps reflecting spell access at that level",
-         %{system: system, character: character} do
-      # XP for level 5 (6500)
-      xp_effect = %Effect{target: {"character_trait", "experience_points", "total"}, value: 6500}
-      character = %{character | effects: [xp_effect]}
-
-      slots = Characters.compute_pending_choice_slots(system, character)
-      spells_known = Enum.filter(slots, &(&1.progression_id == "spells_known"))
-
-      # At level 5, a sorcerer knows 6 spells total; all 6 slots are pending
-      assert length(spells_known) == 6
-
-      # Each slot's cap should reflect the max spell level available when that slot was earned:
-      # levels 1-2 cap at 1, levels 3-4 cap at 2, level 5 caps at 3
-      caps = Enum.map(spells_known, & &1.max_level_cap)
-      assert caps == [1, 1, 1, 2, 2, 3]
-    end
-
-    test "already-decided slots are excluded from result",
-         %{system: system, character: character} do
-      # Pre-make one spell decision
-      spell_decision = %Decision{
-        scope: {"character_progression", "spells_known"},
-        choice: "choice_1",
-        selection: "fire_bolt"
-      }
-
-      character = %{character | decisions: character.decisions ++ [spell_decision]}
-      slots = Characters.compute_pending_choice_slots(system, character)
-      spells_known = Enum.filter(slots, &(&1.progression_id == "spells_known"))
-
-      # 2 total at level 1, minus 1 decided = 1 remaining
-      assert length(spells_known) == 1
-    end
-
-    test "returns empty list when no level-capped progressions exist",
-         %{system: system, character: character} do
-      # Remove all character_progression metadata that has max_level_node
-      concept_metadata =
-        Map.reject(system.concept_metadata, fn {{type, _id}, meta} ->
-          type == "character_progression" and get_in(meta, ["filter", "max_level_node"]) != nil
-        end)
-
-      system = %{system | concept_metadata: concept_metadata}
-      assert Characters.compute_pending_choice_slots(system, character) == []
-    end
-  end
-
-  describe "concept_roll!/4" do
-    setup do
-      system = RuleSystems.load_system!("dnd_5e_srd")
-      attrs = ~w[strength dexterity constitution wisdom intelligence charisma]
-      generated = Map.new(attrs, &{{"ability", &1, "base_score"}, 10})
-
-      character = %Character{
-        name: "Test Character",
-        generated_values: generated,
-        effects: [],
-        decisions: [],
-        metadata: %ExTTRPGDev.Characters.Metadata{
-          slug: "test_roll_char",
-          rule_system: "dnd_5e_srd"
-        }
-      }
-
-      {:ok, system: system, character: character}
-    end
-
-    test "returns a valid result for a known concept", %{system: system, character: character} do
-      result = Characters.concept_roll!(system, character, "skill", "acrobatics")
-
-      assert Enum.sum(result.rolls) in 1..20
-      assert result.bonus == 0
-      assert result.total == Enum.sum(result.rolls) + result.bonus
-      assert result.type_id == "skill"
-      assert result.concept_id == "acrobatics"
-      assert result.dice == "1d20"
-    end
-
-    test "raises when no roll is defined for the concept type", %{
-      system: system,
-      character: character
-    } do
-      assert_raise RuntimeError, ~r/No roll defined/, fn ->
-        Characters.concept_roll!(system, character, "ability", "strength")
-      end
-    end
-
-    test "raises for an unknown concept", %{system: system, character: character} do
-      assert_raise RuntimeError, ~r/not found/, fn ->
-        Characters.concept_roll!(system, character, "skill", "not_a_real_skill")
-      end
-    end
-  end
-
-  describe "requires filtering in concept_options/4" do
-    setup do
-      {:ok, system: RuleSystems.load_system!("dnd_5e_srd")}
-    end
-
-    defp feat_options(system, str_score) do
-      meta = %{"type" => "feat", "required_count" => "1"}
-      resolved = %{{"ability", "strength", "total_score"} => str_score}
-      Characters.concept_options(meta, system.concept_metadata, MapSet.new(), resolved)
-    end
-
-    test "grappler requires STR >= 13 and ability_score_improvement is always available",
-         %{system: system} do
-      below = feat_options(system, 12)
-      refute "grappler" in below
-      assert "ability_score_improvement" in below
-
-      assert "grappler" in feat_options(system, 13)
-    end
-  end
-
-  describe "pending sub-choices from selected progression concepts" do
-    @asi_progression %{
-      "name" => "ASI or Feat",
-      "required_count" => "1",
-      "type" => "feat"
-    }
-
-    @asi_feat_meta %{
-      "name" => "Ability Score Improvement",
-      "choices" => %{
-        "asi_point_1" => %{
-          "type" => "ability",
-          "contributes_field" => "total_score",
-          "contributes_value" => 1
-        },
-        "asi_point_2" => %{
-          "type" => "ability",
-          "contributes_field" => "total_score",
-          "contributes_value" => 1
-        }
-      }
-    }
-
-    @ability_meta %{
-      {"ability", "strength"} => %{"name" => "Strength"},
-      {"ability", "dexterity"} => %{"name" => "Dexterity"}
-    }
-
-    defp asi_system do
-      concept_metadata =
-        Map.merge(
-          %{
-            {"character_progression", "asi_or_feat"} => @asi_progression,
-            {"feat", "ability_score_improvement"} => @asi_feat_meta
-          },
-          @ability_meta
-        )
-
-      %ExTTRPGDev.RuleSystems.LoadedSystem{
-        module: nil,
-        graph: nil,
-        nodes: %{},
-        rolling_methods: %{},
-        effects: [],
-        concept_metadata: concept_metadata
-      }
-    end
-
-    defp asi_character(decisions) do
-      %Character{
-        name: "Test",
-        generated_values: %{},
-        effects: [],
-        decisions: decisions,
-        metadata: %ExTTRPGDev.Characters.Metadata{slug: "test", rule_system: "dnd_5e_srd"}
-      }
-    end
-
-    test "no sub-choices when feat has not been selected" do
-      system = asi_system()
-      choices = Characters.pending_choices(system, asi_character([]), %{})
-
-      sub = Enum.filter(choices, &Map.has_key?(&1, :scope_type))
-      assert sub == []
-    end
-
-    defp asi_sub_choices do
-      decisions = [
-        %Decision{
-          scope: {"character_progression", "asi_or_feat"},
-          choice: "choice_1",
-          selection: "ability_score_improvement"
-        }
-      ]
-
-      choices = Characters.pending_choices(asi_system(), asi_character(decisions), %{})
-      Enum.filter(choices, &Map.has_key?(&1, :scope_type))
-    end
-
-    test "two sub-choices appear after selecting ability_score_improvement" do
-      sub = asi_sub_choices()
-      assert length(sub) == 2
-      assert Enum.map(sub, & &1.id) |> Enum.sort() == ["asi_point_1", "asi_point_2"]
-    end
-
-    test "sub-choices carry correct scope and options" do
-      [c | _] = asi_sub_choices()
-      assert c.scope_type == "feat"
-      assert c.scope_id == "ability_score_improvement"
-      assert Enum.sort(c.options) == ["dexterity", "strength"]
-    end
-
-    test "sub-choice count decreases as decisions are resolved" do
-      decisions = [
-        %Decision{
-          scope: {"character_progression", "asi_or_feat"},
-          choice: "choice_1",
-          selection: "ability_score_improvement"
-        },
-        %Decision{
-          scope: {"feat", "ability_score_improvement"},
-          choice: "asi_point_1",
-          selection: "strength"
-        }
-      ]
-
-      choices = Characters.pending_choices(asi_system(), asi_character(decisions), %{})
-
-      sub = Enum.filter(choices, &Map.has_key?(&1, :scope_type))
-      assert length(sub) == 1
-      assert hd(sub).id == "asi_point_2"
-    end
-
-    test "no sub-choices remain after all points resolved" do
-      decisions = [
-        %Decision{
-          scope: {"character_progression", "asi_or_feat"},
-          choice: "choice_1",
-          selection: "ability_score_improvement"
-        },
-        %Decision{
-          scope: {"feat", "ability_score_improvement"},
-          choice: "asi_point_1",
-          selection: "strength"
-        },
-        %Decision{
-          scope: {"feat", "ability_score_improvement"},
-          choice: "asi_point_2",
-          selection: "dexterity"
-        }
-      ]
-
-      choices = Characters.pending_choices(asi_system(), asi_character(decisions), %{})
-
-      sub = Enum.filter(choices, &Map.has_key?(&1, :scope_type))
-      assert sub == []
-    end
-  end
-
-  describe "auto_resolve_pending/2" do
-    setup do
-      system = RuleSystems.load_system!("dnd_5e_srd")
-      decisions = Characters.random_decisions(system)
-      character = Character.gen_character!(system, decisions)
-      slots = Characters.compute_pending_choice_slots(system, character)
-      character = %{character | pending_choice_slots: slots}
-      %{system: system, character: character}
-    end
-
-    test "clears all level-1 pending choices for a fresh character", %{
-      system: system,
-      character: character
-    } do
-      resolved_character = Characters.auto_resolve_pending(system, character)
-
-      all_effects = Characters.active_effects(system, resolved_character)
-
-      resolved =
-        ExTTRPGDev.RuleSystem.Evaluator.evaluate!(
-          system,
-          resolved_character.generated_values,
-          all_effects
-        )
-
-      choices = Characters.pending_choices(system, resolved_character, resolved)
-
-      level_1_pending =
-        Enum.filter(choices, fn c ->
-          c.type == :pending and c[:earned_at_level] in [nil, 1]
-        end)
-
-      assert level_1_pending == []
-    end
-
-    test "is a no-op when called again after all level-1 choices are resolved", %{
-      system: system,
-      character: character
-    } do
-      resolved_once = Characters.auto_resolve_pending(system, character)
-      resolved_twice = Characters.auto_resolve_pending(system, resolved_once)
-      assert resolved_once == resolved_twice
-    end
-
-    test "selection progressions add decisions pointing to valid concepts", %{
-      system: system,
-      character: character
-    } do
-      resolved_character = Characters.auto_resolve_pending(system, character)
-
-      resolved_character.decisions
-      |> Enum.filter(fn d -> match?({"character_progression", _}, d.scope) end)
-      |> Enum.each(fn decision ->
-        {"character_progression", prog_id} = decision.scope
-        meta = system.concept_metadata[{"character_progression", prog_id}]
-
-        if meta && Map.has_key?(meta, "type") do
-          concept_type = meta["type"]
-
-          assert system.concept_metadata[{concept_type, decision.selection}] != nil,
-                 "selection #{inspect(decision.selection)} not found in #{concept_type} concepts"
-        end
-      end)
-    end
-  end
-
-  describe "random_resolve_all/2" do
-    setup do
-      system = RuleSystems.load_system!("dnd_5e_srd")
-      decisions = Characters.random_decisions(system)
-      character = Character.gen_character!(system, decisions)
-      slots = Characters.compute_pending_choice_slots(system, character)
-
-      character =
-        Characters.auto_resolve_pending(system, %{character | pending_choice_slots: slots})
-
-      # Award XP to level 2 to generate HP pending choices
-      xp_target = {"character_trait", "experience_points", "total"}
-
-      character = %{
-        character
-        | effects: character.effects ++ [%Effect{target: xp_target, value: 300}]
-      }
-
-      slots = Characters.compute_pending_choice_slots(system, character)
-      character = %{character | pending_choice_slots: slots}
-
-      %{system: system, character: character}
-    end
-
-    test "returns updated character with no pending choices", %{
-      system: system,
-      character: character
-    } do
-      {resolved, _resolutions} = Characters.random_resolve_all(system, character)
-
-      all_effects = Characters.active_effects(system, resolved)
-
-      resolved_map =
-        ExTTRPGDev.RuleSystem.Evaluator.evaluate!(system, resolved.generated_values, all_effects)
-
-      choices = Characters.pending_choices(system, resolved, resolved_map)
-
-      assert Enum.filter(choices, &(&1.type == :pending)) == []
-    end
-
-    test "returns a resolution entry for each resolved choice", %{
-      system: system,
-      character: character
-    } do
-      all_effects = Characters.active_effects(system, character)
-
-      resolved_map =
-        ExTTRPGDev.RuleSystem.Evaluator.evaluate!(system, character.generated_values, all_effects)
-
-      pending_before =
-        Enum.filter(
-          Characters.pending_choices(system, character, resolved_map),
-          &(&1.type == :pending)
-        )
-
-      {_resolved, resolutions} = Characters.random_resolve_all(system, character)
-
-      assert length(resolutions) >= length(pending_before)
-    end
-
-    test "selection resolutions carry a concept_type and selection_id", %{
-      system: system,
-      character: character
-    } do
-      {_resolved, resolutions} = Characters.random_resolve_all(system, character)
-
-      selection_resolutions = Enum.filter(resolutions, &(&1.selection_id != nil))
-
-      Enum.each(selection_resolutions, fn r ->
-        assert r.concept_type != nil
-        assert system.concept_metadata[{r.concept_type, r.selection_id}] != nil
-      end)
-    end
-
-    test "value resolutions carry a rolled_value and method", %{
-      system: system,
-      character: character
-    } do
-      {_resolved, resolutions} = Characters.random_resolve_all(system, character)
-
-      value_resolutions = Enum.filter(resolutions, &(&1.rolled_value != nil))
-
-      Enum.each(value_resolutions, fn r ->
-        assert is_integer(r.rolled_value) and r.rolled_value > 0
-        assert r.method in ["rolled", "average"]
-      end)
-    end
-  end
-
-  describe "activate/4" do
-    defp spell_inv_rules do
-      {:ok, rules} =
-        InventoryRules.from_map(%{
-          "inventory_type" => %{
-            "equipment" => %{
-              "activation_field" => "equipped",
-              "schema" => %{"equipped" => %{"type" => "boolean", "default" => false}}
-            },
-            "spell" => %{
-              "activation_field" => "prepared",
-              "activate_command" => "prepare",
-              "schema" => %{"prepared" => %{"type" => "boolean", "default" => false}},
-              "add_on_progression" => [
-                %{"progression" => "cantrips", "auto_activate" => true},
-                %{"progression" => "spells_known"}
-              ],
-              "preparation" => %{
-                "mode_field" => "preparation_mode",
-                "activation_mode" => "prepared",
-                "pool_field" => "preparation_pool",
-                "cap_field" => "preparation_cap",
-                "level_field" => "level",
-                "max_level_node" => ["character_trait", "max_spell_level", "level"],
-                "always_prepared" => %{"metadata_key" => "always_prepared"},
-                "auto_activate_when" => %{
-                  "class_field" => "preparation_mode",
-                  "class_value" => "all"
-                },
-                "pool" => %{
-                  "spellbook" => %{
-                    "scope_type" => "character_progression",
-                    "scope_id" => "spells_known",
-                    "management" => "toggle_field"
-                  },
-                  "class_spells" => %{
-                    "class_filter_field" => "classes",
-                    "management" => "add_remove"
-                  }
-                }
-              }
-            }
-          }
-        })
-
-      rules
-    end
-
-    defp spell_system(concept_metadata \\ %{}) do
-      %LoadedSystem{
-        module: %{character_building_choices: [%{concept_type: "class"}]},
-        graph: nil,
-        nodes: %{},
-        rolling_methods: %{},
-        effects: [],
-        concept_metadata: concept_metadata,
-        inventory_rules: spell_inv_rules()
-      }
-    end
-
-    test "returns error for unknown inventory type" do
-      assert {:error, {:unknown_inventory_type, "weapon"}} =
-               Characters.activate(spell_system(), minimal_character([]), "weapon", [])
-    end
-
-    test "returns error for non-preparation type" do
-      assert {:error, {:not_a_preparation_type, "equipment"}} =
-               Characters.activate(spell_system(), minimal_character([]), "equipment", [])
-    end
-
-    test "returns error when no class decision has preparation mode" do
-      assert {:error, :no_preparation_class} =
-               Characters.activate(spell_system(), minimal_character([]), "spell", [])
-    end
-
-    test "returns error when class preparation mode is not prepared" do
-      meta = %{{"class", "bard"} => %{"preparation_mode" => "all"}}
-      character = minimal_character([%Decision{scope: nil, choice: "class", selection: "bard"}])
-
-      assert {:error, {:mode_not_prepared, "all"}} =
-               Characters.activate(spell_system(meta), character, "spell", [])
-    end
-
-    test "toggle_field: sets prepared true for listed items and false for others" do
-      nodes = %{
-        {"class", "wizard", "preparation_cap"} => %Node{type: :accumulator, base: "3"},
-        {"character_trait", "max_spell_level", "level"} => %Node{type: :accumulator, base: "2"}
-      }
-
-      concept_metadata = %{
-        {"class", "wizard"} => %{
-          "preparation_mode" => "prepared",
-          "preparation_pool" => "spellbook"
-        },
-        {"spell", "fire_bolt"} => %{"level" => 1},
-        {"spell", "cure_wounds"} => %{"level" => 1}
-      }
-
-      {:ok, built} =
-        ExTTRPGDev.RuleSystem.Graph.build(%{
-          nodes: nodes,
-          effects: [],
-          concept_metadata: concept_metadata,
-          rolling_methods: %{}
-        })
-
-      system = %LoadedSystem{
-        module: %{character_building_choices: [%{concept_type: "class"}]},
-        graph: built.graph,
-        nodes: built.nodes,
-        rolling_methods: %{},
-        effects: [],
-        concept_metadata: concept_metadata,
-        inventory_rules: spell_inv_rules()
-      }
-
-      decisions = [
-        %Decision{scope: nil, choice: "class", selection: "wizard"},
-        %Decision{
-          scope: {"character_progression", "spells_known"},
-          choice: "choice_1",
-          selection: "fire_bolt"
-        },
-        %Decision{
-          scope: {"character_progression", "spells_known"},
-          choice: "choice_2",
-          selection: "cure_wounds"
-        }
-      ]
-
-      inventory = [
-        %InventoryItem{
-          concept_type: "spell",
-          concept_id: "fire_bolt",
-          fields: %{"prepared" => false}
-        },
-        %InventoryItem{
-          concept_type: "spell",
-          concept_id: "cure_wounds",
-          fields: %{"prepared" => false}
-        }
-      ]
-
-      character = %{minimal_character(decisions) | inventory: inventory}
-
-      assert {:ok, updated} = Characters.activate(system, character, "spell", ["fire_bolt"])
-
-      assert Enum.find(updated.inventory, &(&1.concept_id == "fire_bolt")).fields["prepared"] ==
-               true
-
-      assert Enum.find(updated.inventory, &(&1.concept_id == "cure_wounds")).fields["prepared"] ==
-               false
-    end
-
-    test "toggle_field: preserves cantrips (outside eligible pool) when preparing leveled spells" do
-      nodes = %{
-        {"class", "wizard", "preparation_cap"} => %Node{type: :accumulator, base: "3"},
-        {"character_trait", "max_spell_level", "level"} => %Node{type: :accumulator, base: "2"}
-      }
-
-      concept_metadata = %{
-        {"class", "wizard"} => %{
-          "preparation_mode" => "prepared",
-          "preparation_pool" => "spellbook"
-        },
-        {"spell", "prestidigitation"} => %{"level" => 0},
-        {"spell", "magic_missile"} => %{"level" => 1}
-      }
-
-      {:ok, built} =
-        ExTTRPGDev.RuleSystem.Graph.build(%{
-          nodes: nodes,
-          effects: [],
-          concept_metadata: concept_metadata,
-          rolling_methods: %{}
-        })
-
-      system = %LoadedSystem{
-        module: %{character_building_choices: [%{concept_type: "class"}]},
-        graph: built.graph,
-        nodes: built.nodes,
-        rolling_methods: %{},
-        effects: [],
-        concept_metadata: concept_metadata,
-        inventory_rules: spell_inv_rules()
-      }
-
-      decisions = [
-        %Decision{scope: nil, choice: "class", selection: "wizard"},
-        %Decision{
-          scope: {"character_progression", "spells_known"},
-          choice: "choice_1",
-          selection: "magic_missile"
-        }
-      ]
-
-      inventory = [
-        %InventoryItem{
-          concept_type: "spell",
-          concept_id: "prestidigitation",
-          fields: %{"prepared" => true}
-        },
-        %InventoryItem{
-          concept_type: "spell",
-          concept_id: "magic_missile",
-          fields: %{"prepared" => false}
-        }
-      ]
-
-      character = %{minimal_character(decisions) | inventory: inventory}
-
-      assert {:ok, updated} = Characters.activate(system, character, "spell", ["magic_missile"])
-
-      assert Enum.find(updated.inventory, &(&1.concept_id == "magic_missile")).fields["prepared"] ==
-               true
-
-      assert Enum.find(updated.inventory, &(&1.concept_id == "prestidigitation")).fields[
-               "prepared"
-             ] == true
-    end
-
-    test "add_remove: preserves cantrips when preparing leveled spells" do
-      nodes = %{
-        {"class", "cleric", "preparation_cap"} => %Node{type: :accumulator, base: "4"},
-        {"character_trait", "max_spell_level", "level"} => %Node{type: :accumulator, base: "2"}
-      }
-
-      concept_metadata = %{
-        {"class", "cleric"} => %{
-          "preparation_mode" => "prepared",
-          "preparation_pool" => "class_spells"
-        },
-        {"spell", "sacred_flame"} => %{"level" => 0, "classes" => ["cleric"]},
-        {"spell", "bless"} => %{"level" => 1, "classes" => ["cleric"]},
-        {"spell", "cure_wounds"} => %{"level" => 1, "classes" => ["cleric"]}
-      }
-
-      {:ok, built} =
-        ExTTRPGDev.RuleSystem.Graph.build(%{
-          nodes: nodes,
-          effects: [],
-          concept_metadata: concept_metadata,
-          rolling_methods: %{}
-        })
-
-      system = %LoadedSystem{
-        module: %{character_building_choices: [%{concept_type: "class"}]},
-        graph: built.graph,
-        nodes: built.nodes,
-        rolling_methods: %{},
-        effects: [],
-        concept_metadata: concept_metadata,
-        inventory_rules: spell_inv_rules()
-      }
-
-      decisions = [%Decision{scope: nil, choice: "class", selection: "cleric"}]
-
-      inventory = [
-        %InventoryItem{
-          concept_type: "spell",
-          concept_id: "sacred_flame",
-          fields: %{"prepared" => true}
-        }
-      ]
-
-      character = %{minimal_character(decisions) | inventory: inventory}
-
-      assert {:ok, updated} = Characters.activate(system, character, "spell", ["bless"])
-
-      spell_inventory = Enum.filter(updated.inventory, &(&1.concept_type == "spell"))
-
-      # sacred_flame (cantrip) preserved; bless added; cure_wounds (eligible, not requested) absent
-      assert MapSet.new(spell_inventory, & &1.concept_id) == MapSet.new(["sacred_flame", "bless"])
-      assert Enum.all?(spell_inventory, &(&1.fields["prepared"] == true))
-    end
-  end
-
-  describe "add_to_typed_inventory/4" do
-    test "returns character unchanged when progression has no inventory type" do
-      {:ok, inv_rules} = InventoryRules.from_map(%{})
-
-      system = %LoadedSystem{
-        module: nil,
-        graph: nil,
-        nodes: %{},
-        rolling_methods: %{},
-        effects: [],
-        concept_metadata: %{},
-        inventory_rules: inv_rules
-      }
-
-      character = minimal_character([])
-
-      assert {:ok, ^character} =
-               Characters.add_to_typed_inventory(system, character, "spells_known", "fire_bolt")
-    end
-
-    test "determines initial activation from progression config and class condition" do
-      meta = %{{"class", "bard"} => %{"preparation_mode" => "all"}}
-      bard_char = minimal_character([%Decision{scope: nil, choice: "class", selection: "bard"}])
-
-      {:ok, cantrip_char} =
-        Characters.add_to_typed_inventory(
-          spell_system(),
-          minimal_character([]),
-          "cantrips",
-          "fire_bolt"
-        )
-
-      {:ok, spell_char} =
-        Characters.add_to_typed_inventory(
-          spell_system(),
-          minimal_character([]),
-          "spells_known",
-          "cure_wounds"
-        )
-
-      {:ok, bard_spell_char} =
-        Characters.add_to_typed_inventory(
-          spell_system(meta),
-          bard_char,
-          "spells_known",
-          "vicious_mockery"
-        )
-
-      assert hd(cantrip_char.inventory).fields["prepared"] == true
-      assert hd(spell_char.inventory).fields["prepared"] == false
-      assert hd(bard_spell_char.inventory).fields["prepared"] == true
-    end
-  end
-
-  describe "preparation_state/3" do
-    test "returns error for unknown inventory type" do
-      assert {:error, {:unknown_inventory_type, "weapon"}} =
-               Characters.preparation_state(spell_system(), minimal_character([]), "weapon")
-    end
-
-    test "returns mode nil when character has no class with preparation config" do
-      assert {:ok, %{mode: nil}} =
-               Characters.preparation_state(spell_system(), minimal_character([]), "spell")
-    end
-
-    test "returns full preparation state for a character with a prepared-mode class" do
-      nodes = %{
-        {"class", "wizard", "preparation_cap"} => %Node{type: :accumulator, base: "3"},
-        {"character_trait", "max_spell_level", "level"} => %Node{type: :accumulator, base: "2"}
-      }
-
-      concept_metadata = %{
-        {"class", "wizard"} => %{
-          "preparation_mode" => "prepared",
-          "preparation_pool" => "spellbook"
-        },
-        {"spell", "fire_bolt"} => %{"level" => 1}
-      }
-
-      {:ok, built} =
-        ExTTRPGDev.RuleSystem.Graph.build(%{
-          nodes: nodes,
-          effects: [],
-          concept_metadata: concept_metadata,
-          rolling_methods: %{}
-        })
-
-      system = %LoadedSystem{
-        module: %{character_building_choices: [%{concept_type: "class"}]},
-        graph: built.graph,
-        nodes: built.nodes,
-        rolling_methods: %{},
-        effects: [],
-        concept_metadata: concept_metadata,
-        inventory_rules: spell_inv_rules()
-      }
-
-      decisions = [
-        %Decision{scope: nil, choice: "class", selection: "wizard"},
-        %Decision{
-          scope: {"character_progression", "spells_known"},
-          choice: "c1",
-          selection: "fire_bolt"
-        }
-      ]
-
-      inventory = [
-        %InventoryItem{
-          concept_type: "spell",
-          concept_id: "fire_bolt",
-          fields: %{"prepared" => true}
-        }
-      ]
-
-      character = %{minimal_character(decisions) | inventory: inventory}
-
-      assert {:ok, %{mode: "prepared", cap: 3, prepared: ["fire_bolt"]}} =
-               Characters.preparation_state(system, character, "spell")
-    end
-
-    test "always_prepared spells from subclass appear and are filtered by max spell level" do
-      nodes = %{
-        {"class", "cleric", "preparation_cap"} => %Node{type: :accumulator, base: "4"},
-        {"character_trait", "max_spell_level", "level"} => %Node{type: :accumulator, base: "1"}
-      }
-
-      concept_metadata = %{
-        {"class", "cleric"} => %{
-          "preparation_mode" => "prepared",
-          "preparation_pool" => "class_spells",
-          "choices" => %{"subclass" => %{"type" => "class"}}
-        },
-        {"class", "life_domain"} => %{
-          "always_prepared" => ["bless", "cure_wounds", "hold_person"]
-        },
-        {"spell", "bless"} => %{"level" => 1, "classes" => ["cleric"]},
-        {"spell", "cure_wounds"} => %{"level" => 1, "classes" => ["cleric"]},
-        {"spell", "hold_person"} => %{"level" => 2, "classes" => ["cleric"]}
-      }
-
-      {:ok, built} =
-        ExTTRPGDev.RuleSystem.Graph.build(%{
-          nodes: nodes,
-          effects: [],
-          concept_metadata: concept_metadata,
-          rolling_methods: %{}
-        })
-
-      system = %LoadedSystem{
-        module: %{character_building_choices: [%{concept_type: "class"}]},
-        graph: built.graph,
-        nodes: built.nodes,
-        rolling_methods: %{},
-        effects: [],
-        concept_metadata: concept_metadata,
-        inventory_rules: spell_inv_rules()
-      }
-
-      decisions = [
-        %Decision{scope: nil, choice: "class", selection: "cleric"},
-        %Decision{scope: {"class", "cleric"}, choice: "subclass", selection: "life_domain"}
-      ]
-
-      character = minimal_character(decisions)
-
-      assert {:ok, state} = Characters.preparation_state(system, character, "spell")
-      assert state.mode == "prepared"
-      # bless and cure_wounds are within max_spell_level 1; hold_person (level 2) is filtered out
-      assert Enum.sort(state.always_prepared) == ["bless", "cure_wounds"]
-    end
-
-    test "always_prepared accumulates from all active concepts, not just subclass" do
-      nodes = %{
-        {"class", "cleric", "preparation_cap"} => %Node{type: :accumulator, base: "4"},
-        {"character_trait", "max_spell_level", "level"} => %Node{type: :accumulator, base: "2"}
-      }
-
-      concept_metadata = %{
-        {"class", "cleric"} => %{
-          "preparation_mode" => "prepared",
-          "preparation_pool" => "class_spells",
-          "choices" => %{"subclass" => %{"type" => "class"}}
-        },
-        # subclass contributes one spell
-        {"class", "life_domain"} => %{"always_prepared" => ["bless"]},
-        # race contributes a different spell via the same metadata key
-        {"race", "aasimar"} => %{"always_prepared" => ["guiding_bolt"]},
-        {"spell", "bless"} => %{"level" => 1, "classes" => ["cleric"]},
-        {"spell", "guiding_bolt"} => %{"level" => 1, "classes" => ["cleric"]}
-      }
-
-      {:ok, built} =
-        ExTTRPGDev.RuleSystem.Graph.build(%{
-          nodes: nodes,
-          effects: [],
-          concept_metadata: concept_metadata,
-          rolling_methods: %{}
-        })
-
-      system = %LoadedSystem{
-        module: %{character_building_choices: [%{concept_type: "class"}, %{concept_type: "race"}]},
-        graph: built.graph,
-        nodes: built.nodes,
-        rolling_methods: %{},
-        effects: [],
-        concept_metadata: concept_metadata,
-        inventory_rules: spell_inv_rules()
-      }
-
-      decisions = [
-        %Decision{scope: nil, choice: "class", selection: "cleric"},
-        %Decision{scope: {"class", "cleric"}, choice: "subclass", selection: "life_domain"},
-        %Decision{scope: nil, choice: "race", selection: "aasimar"}
-      ]
-
-      character = minimal_character(decisions)
-
-      assert {:ok, state} = Characters.preparation_state(system, character, "spell")
-      assert Enum.sort(state.always_prepared) == ["bless", "guiding_bolt"]
-    end
-  end
 end
